Problem: Env struct has grown too large
Solution: extract it into its own submodule
